Implement organya fixes*
It's still is a mystery whether all of this is more or less correct than
how it previously was; while that was based on how Orgmaker 2 works,
this is apparently based on Cave Story proper.

The values are near enough though.

Also implemented a method to fake each channel having two-shot voices.